Add donate command.
This bot is old as shit and not really maintained anymore. Might re-code it later and add a ton of fun shit but for now people can donate if they want. That might make me re-code it and make it better which could be fun but I don't really have time.